Last Commit
Ran into a bug which is apparently related to DirectX from what I can
tell. Can't actually run the game now, but i've added in a text box
which displays the amount of time the player has survived and also a
group of monsters which spawn as a wave. They have hit detection,
however I can't test much due to the error.